changed card and deck: / changed signature of functions that are getters to be const / Card::equals is now const / changed unsigned int to uint32_t - cards and decks can no longer be copy-constructed - cards no longer have a standard constructor
/ fixed issue with the shuffler class, where Shuffler::shuffleDeck was protected

+ added the ShuffleEngine

+++ CardGame/CutShuffle.cpp
@@ -1,6 +1,6 @@
 #include "CutShuffle.h"
 #define CUT_RANGE 5
-#define RAND_INT RandomGenerator::getInstance()->get() * CUT_RANGE + 1
+#define RAND_INT int(RandomGenerator::getInstance()->get() * CUT_RANGE) + 1
 
 CutShuffle::CutShuffle() : Shuffler(10) { }
 CutShuffle::CutShuffle(int cycles) : Shuffler(cycles) { }
